Add the `bundle` command to create PR's with gem updates
Might be worth improving the command naming scheme to allow for `bundle:update` or  `bundle-update` instead to be a bit more clear. But wasn't in scope for this lab.

In mndx we allowed the command to update the gems in all repos at once but since I'm assuming that would not currently work on the `mynewsdesk` repo I'm forcing explicit listing of the repos to update.